requirements.txt
altgraph==0.17.2
fdb==2.0.2
future==0.18.2
importlib-metadata==4.11.3
pefile==2021.9.3
pyinstaller==4.10
pyinstaller-hooks-contrib==2022.2
python-dotenv==0.19.2
pywin32-ctypes==0.2.0
typing_extensions==4.1.1
zipp==3.7.0
